Added HTTP error handling.
Errors that occur from HTTP break the script. I've added the on error handler and pulled through the HTTP error to then write to the result. The callback has also been fixed to handle correct "error" and "result/data" positions if the error callback was to be used in the future.